requirements.txt
Flask==3.1.2
Flask-CORS==4.0.0
Flask-APScheduler==1.13.1
Flask-Mail==0.10.0
Flask-WTF==1.2.2
google-generativeai>=0.8.5
python-dotenv>=1.1.1
Werkzeug==3.1.3
